Moves the basic yoast-components to the components package
