+++ packages/components/Paper.js
@@ -1,7 +1,7 @@
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import colors from "../../style-guide/colors.json";
+import colors from "../yoast-components/style-guide/colors.json";
 
 /**
  * Returns a Div with a paper style.
