+++ packages/components/YoastLogo.js
