+++ packages/components/YoastSeoIcon.js
