+++ packages/components/composites/Table/Cell.js
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import colors from "../../../style-guide/colors.json";
+import colors from "../../../yoast-components/style-guide/colors.json";
 
 /**
  * Separatifies an element.
