+++ packages/components/composites/Table/Row.js
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import colors from "../../../style-guide/colors.json";
+import colors from "../../../yoast-components/style-guide/colors.json";
 
 // The Rows are flex containers.
 export const Row = styled.li`
